Move components to module folder
