+++ src/modules/Logout.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import 'bootstrap/dist/css/bootstrap.min.css';
 import Button from 'react-bootstrap/Button';
-import './Login.css';
+import '../Login.css';
 import { useAuth0 } from '@auth0/auth0-react';
 
 function Logout() {
